Try again
I have two duplicate directories of VBA files. Reason being is the name. I thought by uploading a directory with files "VisualBasic" it would override the existing and therefore indicate the 'diffs' as it normally does. The problem was the folder I thought I was replacing was named VBA Modules. Ok so now I'm going to try again. If it doesn't work, I'll have three copies of the same. Yikes. And on the web I don't even think you can delete them. I'll have to find a different way to delete them.