Newer and better
significant refactor. The Pivot Table maker now let's you custom delete rows (records) in the format tab that don't apply to your job (e.g. admin time, etc). It will only run the data validator sub if the PTFormat tab doesn't exist (changes a bool), so you have to run it twice to get the outcome but that's alright. For now, maybe we can optimize. Also, you don't have to be on the report worksheet to run it as it will only run the FormatTable conditionally, if the PTFormat tab doesn't work. In other words, more user friendly with higher functionality. 

to consider doing- if run with pivot table worksheet already existing, delete the table and run again based on the new table data.