Works with data validator but more complex
The increasing complexity worries me, as things could break. Also formatting has to be exact so user has to be trained. User interface (messages boxes) could be updated to be smoother. 

User has to make pivot table format equal or extra worksheets will be created duplicating the values. Ways to improve this, for another day.