Gloriously working
Done, finally, free of all bugs. Dynamic template selection - selects template if AdHocTemplateStr is "" (so once) and if so, selects from the first worksheet found (taking advantage of the search for non existent worksheets. We presume that is a good worksheet (to use as a template) and define that variable string, so the code is very dynamic and robust. 

It works and it works pretty quickly. It is optimized to stop searching through cells on the pivot table when two hits are found (because doctors have new patients and old patients so there will only be two instances) so this helps to optimize things for speed and performance. 

There may be ways to optimize better. There are still some theoretical holes (where there can be number errors) but they require exceptional cases of bad naming not covered by the two identified name anomalies.

The standard way to name a doctor worksheet is F. Last, or first initial, dot, space, last name. 
The two anomalies are no space and where a name has a hyphen on the pivot table, to truncate that. So  we have handled these exceptions in the code through a function call with ByRef keyword. (Could have used a collection perhaps). 

So if different errors come through in terms of data mismatching (should be rare) the dialogue message box should cover this. 
Example of a rare error- a name both without the space and a truncation of hyphenated last name. Or other such thing. Not super worried. 

Good robust code but can always find ways to improve. 

Can replace some source code in the main and elsewhere with function calls for cleanliness but pretty good.