OneStop.xlam updated
this is where effectively all my macros are anyway. the rest is just for granularity and clutter. 

On my TODO list is to organize and refactor them, so there are not so many modules and the modules that make sense can be organized and tidied and found. Then I can clean up what's on the ribbon later. I hate how changing names of modules screws up ribbon links. 

Made a few new macros with Dictionary functions and Regex!