powerful medical ADP csv maker added
Creates a csv file and writes to the debugger (immediate window) in excel VBA for ADP reports for medical JEs. Links department with value for a given search (for example gross wages by department or 'health deduction' by department. User to specify which except in case of gross wages, which is a separate sub.

Makes use of regex, so appropriate library(ies) must be added in tools- reference. google it.  

The deduction sub prompts user for code (e.g. 'H' for health code, 'v' for voluntary life insurance - they should know- part of ADP). V is ambiguous code used for multiple items but this sub screens out the non-relevant ones to return a good value. 

No promise is made that this works for every possible code or the macro doesn't crash but it works for the ones used for our work- namely H, V, O (dental), etc. The goal is to save time and this saves a massive amt of time, but the goal is not to create a perfect automated interface of all potential needs. This works, within the scope it was designed for. 

Also created a copy to clipboard sub based on regular expression as well as a copy value to clipboard (saving the need to right click or whatever).